backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
google-genai>=1.26.0
openai>=1.0.0
python-dotenv==1.0.0
pydantic-settings==2.0.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
firebase-admin==6.4.0
requests==2.31.0
stripe==7.10.0
pandas==2.1.4
python-magic==0.4.27
Pillow==10.0.0
slowapi==0.1.9
aiofiles==23.2.1
bleach==6.1.0
PyYAML==6.0.2
